+++ src/main/java/com/fathi/newrootacademymanager/models/Income.java
@@ -2,6 +2,7 @@ package com.fathi.newrootacademymanager.models;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -11,7 +12,7 @@ public class Income {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
     @Column(nullable = false)
-    private float amount;
+    private BigDecimal amount;
     @Column(nullable = false)
     private String details;
     @Column(name = "create time", nullable = false, columnDefinition = "DATETIME DEFAULT CURRENT_TIMESTAMP")
@@ -24,12 +25,12 @@ public class Income {
 
     public Income() {}
 
-    public Income(float amount, String details) {
+    public Income(BigDecimal amount, String details) {
         this.amount = amount;
         this.details = details;
     }
 
-    public Income(float amount, String details, Student student) {
+    public Income(BigDecimal amount, String details, Student student) {
         this.amount = amount;
         this.details = details;
         this.student = student;
@@ -43,11 +44,11 @@ public class Income {
         this.id = id;
     }
 
-    public float getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
